Add BARGE logo to logos directory.
This is not currently used, but a couple versions use the copy in the
static directory.  This will allow Hugo operators to find the data.
However, it looks like the logo is not that large (~35k) and it's used
on every page, so at least we aren't re-downloading it.  But because the
logo assets are all over the directories for the site, I want a copy
here so I can find it later.